Project Progress
This is probably not ready to be added to the main github project. I
altered what the left click would do- it currently sets everyone but the
sidekick on fire- and I changed the NPC update function.

Nonetheless, it's all in there and, if you open the DonMobFireTest
scene, you can try it out. I couldn't get the walls to work right,
though.